added separated utils for days
